Keep fiddling to try and enable automated tests in dev environment
