+++ lambda_functions/nci_monitoring/dea_monitoring/.graphqlconfig
@@ -1,0 +1,16 @@
+{
+  "name": "GitHub API V4 GraphQL Schema",
+  "schemaPath": "github-schema.graphql",
+  "extensions": {
+    "endpoints": {
+      "GitHub API V4 GraphQL Endpoint": {
+        "url": "https://api.github.com/graphql",
+        "headers": {
+          "Authorization": "Bearer ${env:GITHUB-ACCESS-TOKEN}",
+          "user-agent": "JS GraphQL"
+        },
+        "introspect": false
+      }
+    }
+  }
+}
