+++ lambda_functions/nci_monitoring/serverless.yml
@@ -18,12 +18,14 @@ plugins:
 
 custom:
   pythonRequirements:
-#    pipCmdExtraArgs:
-#      - '--platform'
-#      - 'manylinux1'
-#      - '--only-binary'
-#      - ':all:'
     dockerizePip: non-linux
+  Stage: ${opt:stage, self:provider.stage}
+  profile:
+    dev: dea-dev
+    prod: dea-prod
+  esURL:
+    dev:  Fn::GetAtt: [ ElasticSearchInstance , DomainEndpoint ]
+    prod: search-digitalearthaustralia-lz7w5p3eakto7wrzkmg677yebm.ap-southeast-2.es.amazonaws.com
 
 package:
   exclude:
@@ -38,13 +40,14 @@ package:
 
 resources:
   Description: Monitor Digital Earth Australia quotas and processes running at the NCI
+  Resources: ${self:${opt:stage}Resources}
 
 provider:
   name: aws
   runtime: python3.7
   timeout: 60  # 60 seconds. Default is 6 seconds
   memorySize: 512  # in MB, default is 1024
-  stage: prod
+  stage: dev
   # Since SES is not supported in Sydney region, all emails are pushed to the
   # S3 bucket in us-west-2 region. For s3 to trigger SQS and SQS trigger lambda function execution, all the aws
   # resources need to be in the same region as S3 bucket.
@@ -86,7 +89,7 @@ provider:
     NCI_PROJECTS: 'rs0,v10,u46,fk4,r78,if87'
     NCI_STORAGE: 'gdata1,gdata1a,gdata1b,gdata2,gdata3,short,massdata'
     SSM_USER_PATH: 'orchestrator.raijin.users.default'
-    ES_HOST: search-digitalearthaustralia-lz7w5p3eakto7wrzkmg677yebm.ap-southeast-2.es.amazonaws.com
+    ES_HOST: ${self:custom.esURL.${self:custom.Stage}}
     ES_AWS_REGION: ap-southeast-2
     SSM_GH_TOKEN_PATH: 'orchestrator.github_stats.auth_token'
 
@@ -112,3 +115,28 @@ functions:
           owner: opendatacube
           repo: datacube-core
 
+devResources:
+  Resources:
+    ElasticSearchInstance:
+      Type: AWS::Elasticsearch::Domain
+      Properties:
+        EBSOptions:
+          EBSEnabled: true
+          VolumeType: gp2
+          VolumeSize: 10
+        ElasticsearchClusterConfig:
+          InstanceType: t2.small.elasticsearch
+          InstanceCount: 1
+          DedicatedMasterEnabled: false
+          ZoneAwarenessEnabled: false
+        ElasticsearchVersion: 6.5
+  Outputs:
+    ElasticSearchInstance:
+      Value:
+        Ref: ElasticSearchInstance
+    DomainEndpoint:
+      Value:
+        Fn::GetAtt: [ ElasticSearchInstance, DomainEndpoint ]
+
+
+prodResources:
